Team Leaf Working for Employees Table

+++ src/main/java/northwind/entity/EmployeesRepository.java
@@ -5,10 +5,10 @@ import java.util.List;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EmployeesRepository extends CrudRepository<Employees, Long> {
+public interface EmployeesRepository extends CrudRepository<Employees, Short> {
 
 	List<Employees> findByLastName(String lastName);
-	Employees findByEmployeeID(long employeeID);
+	Employees findByEmployeeID(Short employeeID);
 	
 	@Query(value = "SELECT e FROM Employees e ORDER BY e.employeeID")
 	List<Employees> findAllEmployeesWithPagination();
